vpn_prod/requirements.txt
aiogram>=3.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
loguru>=0.7.2
aiohttp>=3.9.1
py3xui
yookassa
#asyncio
qrcode[pil]
aiocryptopay